+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @idealizesoft-ui/react
 
+## 2.0.3
+
+### Patch Changes
+
+- Password Input
+
 ## 2.0.2
 
 ### Patch Changes