+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @idealizesoft-ui/tokens
 
+## 2.0.1
+
+### Patch Changes
+
+- Password Input
+
 ## 2.0.0
 
 ### Major Changes